+++ react/javascript/src/components/customise/index.ts
@@ -1,0 +1,2 @@
+export * from './CustomRendering'
+export * from './Theme'